+++ packages/ubcli/lib/generateDDL.js
@@ -29,8 +29,9 @@ const _ = require('lodash')
 const fs = require('fs')
 const path = require('path')
 const http = require('http')
-const options = require('@unitybase/base').options
-const argv = require('@unitybase/base').argv
+const base = require('@unitybase/base')
+const options = base.options
+const argv = base.argv
 const execSql = require('./execSql')
 const CRLF = '\n'
 
@@ -44,6 +45,7 @@ module.exports = function generateDDL (cfg) {
       .add({ short: 'cfg', long: 'cfg', param: 'localServerConfig', defaultValue: 'ubConfig.json', searchInEnv: true, help: 'Path to UB server config' })
       .add({ short: 'm', long: 'models', param: 'modelsList', defaultValue: '*', help: 'Comma separated model names for DDL generation. If -e specified this options is ignored' })
       .add({ short: 'e', long: 'entities', param: 'entitiesList', defaultValue: '*', help: 'Comma separated entity names list for DDL generation' })
+      .add({ short: 'c', long: 'connection', param: 'connection', defaultValue: '', help: 'Optional connection name for DDL generation' })
       .add({ short: 'out', long: 'out', param: 'outputPath', defaultValue: process.cwd(), help: 'Folder to output generated DDLs (one file per connection)' })
       .add({ short: 'autorun', long: 'autorun', defaultValue: false, help: 'execute DDL statement after generation. BE CAREFUL! DO NOT USE ON PRODUCTION' })
       .add({ short: 'optimistic', long: 'optimistic', defaultValue: false, help: 'skip errors on execute DDL statement. BE CAREFUL! DO NOT USE ON PRODUCTION' })
@@ -59,7 +61,7 @@ module.exports = function generateDDL (cfg) {
   const session = argv.establishConnectionFromCmdLineAttributes(cfg)
   const conn = session.connection
   try {
-    runDDLGenerator(conn, cfg.autorun, cfg.entities, cfg.models, cfg.out, cfg.optimistic)
+    runDDLGenerator(conn, cfg.autorun, cfg.entities, cfg.models, cfg.out, cfg.optimistic, cfg.connection)
   } finally {
     if (session && session.logout) {
       session.logout()
@@ -74,15 +76,16 @@ module.exports = function generateDDL (cfg) {
  *  @param {String} inModelsCSV
  *  @param {String} outputPath
  *  @param {String} [optimistic=false]
+ *  @param {String} [connectionName]
  *  @private
  */
-function runDDLGenerator (conn, autorun, inEntities, inModelsCSV, outputPath, optimistic) {
+function runDDLGenerator (conn, autorun, inEntities, inModelsCSV, outputPath, optimistic, connectionName) {
   let txtRes
   let entityNames = []
   let inModels = []
 
   const domain = conn.getDomainInfo(true)
-  if (!inEntities && !inModelsCSV) {
+  if (!inEntities && !inModelsCSV && !connectionName) {
     entityNames = Object.keys(domain.entities)
   } else {
     if (inEntities) { // add passed entityNames
@@ -92,6 +95,13 @@ function runDDLGenerator (conn, autorun, inEntities, inModelsCSV, outputPath, op
       inModels = inModelsCSV.split(',')
       domain.eachEntity((entity, entityName) => {
         if (inModels.indexOf(entity.modelName) !== -1) {
+          entityNames.push(entityName)
+        }
+      })
+    }
+    if (connectionName) {
+      domain.eachEntity((entity, entityName) => {
+        if (entity.connectionConfig.name === connectionName) {
           entityNames.push(entityName)
         }
       })
@@ -128,11 +138,35 @@ function runDDLGenerator (conn, autorun, inEntities, inModelsCSV, outputPath, op
       fs.writeFileSync(fileName, txtRes)
       console.log('DDL script is saved to ' + fileName)
       if (autorun) {
-        execSql({
-          connection: connectionName,
-          optimistic: optimistic,
-          file: fileName
-        })
+        if (base.ubVersionNum > 501815) { // execute from file
+          execSql({
+            connection: connectionName,
+            optimistic: optimistic,
+            file: fileName
+          })
+        } else { // for UB < 5.18.16 execute using runSQL endpoint (excSql can fail in case server is started by script)
+          // TODO - remove this code in the middle of 2021
+          let withErrors = false
+          console.log('Run a script ' + fileName)
+          // Many databases (Oracle for example) do not allow to execute several DDL statement in one call
+          for (const part of nonEmptySorted) {
+            for (const stmt of part.statements) {
+              try {
+                if (stmt) {
+                  conn.xhr({ endpoint: 'runSQL', data: stmt, URLParams: { CONNECTION: connectionName } })
+                }
+              } catch (e) {
+                if (!optimistic) {
+                  throw e
+                } else {
+                  console.error(e)
+                  withErrors = true
+                }
+              }
+            }
+          }
+          console.info('Database script', fileName, 'executed', withErrors ? 'with errors!' : 'successfully')
+        }
       }
     } else {
       console.log('Specified entity metadata is congruence with the database for connection ' + connectionName)